requirements.txt
python-dotenv
fuzzywuzzy
pandas
pyspark
mlflow
azure-ai-textanalytics
numpy
boto3
msal
azure-storage-blob
azure-identity
azure-common
msrestazure
functools
requests
regex
fuzzywuzzy